replacement int by long

+++ src/Main.java
@@ -1,8 +1,8 @@
 public class Main {
     public static void main(String[] args) {
-        int[] arr = {5, 7, 4};
+        long[] arr = {5, 7, 4};
         SalesManager salesManager = new SalesManager(arr);
-        int max = salesManager.max();
+        long max = salesManager.max();
         System.out.println(max);
     }
 }
